% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notch_estimator.R
\name{notch_estimator}
\alias{notch_estimator}
\title{Analyzing Bunching at a Notch}
\usage{
notch_estimator(
  earnings,
  zstar,
  t1,
  t2,
  Tax = 0,
  cf_start = NA,
  cf_end = NA,
  exclude_before = NA,
  exclude_after = NA,
  force_after = FALSE,
  binw = 10,
  poly_size = 7,
  convergence = 0.01,
  max_iter = 100,
  select = TRUE,
  draw = TRUE,
  title = "Bunching Visualization",
  varname = "Earnings"
)
}
\arguments{
\item{earnings}{Vector of earnings, hopefully a very large one}

\item{zstar}{Place of notch (critical earning point)}

\item{t1}{Tax rate before notch}

\item{t2}{Tax rate after notch}

\item{Tax}{Lump sum penalty for crossing zstar.}

\item{cf_start}{Number of bins before the notch bin where counter-factual
histogram should start.}

\item{cf_end}{Number of bins after the notch bin where counter-factual
histogram should start.}

\item{exclude_before}{Number of excluded bins before the notch bin.}

\item{exclude_after}{Number of excluded bins after the notch bin.}

\item{force_after}{Should \code{bunch} be forced to use of the provided
\emph{exclude_after} for the end of the bunching, rather than trying to
find the bin where the sum of the integral is zero? See details.}

\item{binw}{Bin width.}

\item{poly_size}{Order of polynomial used to calculate counter-factual
histogram.}

\item{convergence}{Minimal rate of change of bunching estimate to stop
iterations.}

\item{max_iter}{Maximum number of iterations for bunching estimates.}

\item{select}{Should model selection be used to find counter-factual
histogram? See details.}

\item{draw}{Should a graph be drawn?}

\item{title}{Title for plot output}

\item{varname}{Name for running variable, to be displayed in the plot}
}
\value{
\code{notch_estimator } returns a list of the following variables:
\describe{
  \item{\code{e}}{Estimated elasticity}
  \item{\code{Bn}}{The sum of total estimated extra bunching in the area starting
  at cf_start and through the notch bin (zstar) }
  \item{\code{notch_size}}{Distance between notch bin and bin where the estimated
  influence of the notch ends, delta_zed}
  \item{\code{data}}{A data frame with bin mids, counts, counter-factual
  counts, and excluded dummy}
}
}
\description{
Given a notched budget set, this function gets a vector of earnings and
analyzes bunching. This function could be run independently, but best used
through the \code{bunch} function.
}
\details{
By default, \code{notch_estimator} will try to find the end of the notch,
  i.e. a histogram bin where the incentive to bunch wears off. To do this,
  a counter factual distribution is interpolated using the bins inside the
  counter-factual area but outside of the excluded area. This is assumed to be
  the histogram we would have without the tax at the notch point. At the
  observed bunching area, the bins should be much greater than their
  corresponding counter-factual bins. Later on, the bunching should create
  the notch area where the observed bins are lower than the counter-factual.
  factual bin  However, both observed and counter-factual histograms should
  have the same total mass. The end of the notch is determined as the bin
  where the running sum of differences between observed and counter-factual
  bins reaches zero. \code{notch_estimator} goes through an iterative
  process, setting the notch end bin and re-interpolating the counter-factual
  in the notch area using all the bins outside of the notch, trying to find
  a stable right-side boundary.

  A user might want to force a visibly detectable end of notch, rather than
  let \code{notch_estimator} calculate one. Use \code{force_after=TRUE} with
  caution: this sets the notch size as \code{exclude_after} minus
   \code{exclude_before}
  in terms of bins. The notch size is used to calculate elasticity, and
  forcing the wrong notch size might bias the intensive margin elasticity
  estimate. In other settings, e.g. a labor market with extensive margins
  (entry and exit from labor force), forcing the notch size might be helpful.

  For "impure" notches, where the marginal tax rate after the notch is
  different than the one before it, this function disregards the shifting of
  post-notch distribution to the right, as suggested by Kleven (2016).
  Assumption is that the notch effect is much stronger anyway.

  The \code{select} option implements a \code{step} model selection. It runs
  backwards from the full polynomial model (specified by \code{poly_size}),
  trying to find the best explanatory polynomial model for counter-factual
  histogram.
}
\examples{
ability_vec <- 4000 * rbeta(100000, 2, 5)
earning_vec <- sapply(ability_vec, earning_fun, 0.2, 0.2, 0.2, 500, 1000)
bunch_viewer(earning_vec, 1000, 15, 30, 2, 21, binw = 50)
notch_estimator(earning_vec, 1000, 0.2, 0.2, 500, 15, 30, 2, 21, binw = 50,
draw = FALSE)$e

}
\references{
Kleven, H J (2016). \emph{Bunching}, Annual Review of Economics,
  8(1).
}
\seealso{
\code{\link{bunch}}, \code{\link{kink_estimator}}
}
